Move struct ImageOpsExecution to mod image_exec
